Improve Better Victory Screen Compatibility (#23)
Adds a global.finished, because BVS does not set either the two conditions you currently check for. If this is not added then nothing will go wrong because BVS does this internally already. But it will prevent edge cases, for example if BVS is removed after FF victory is achieved, then you'll be able to trigger the victory again.
---------

Co-authored-by: Tom Burrows <tomburrows13@icloud.com>